func_80031874: s32 for a fourth narrow local, census 12 opcodes to 11
`t` held the +0x5AC4 byte and was declared u8 to match the load. s32 removes
the slti and sltiu entries outright -- `(s32)t < 3` on a u8 local was being
done in QImode -- and takes the alignment 248 to 251. That is the fourth local
this session declared to match its load's width and wrong for it.

Installed at -2 over a 280/280 candidate with 238 differences. Both censuses
sum to their length error and this one wins on distinct opcodes and on
alignment, the two keys above the difference count. Flagging the tension
because last tick I went the other way on a one-point alignment gap; three
points plus an opcode is a different call.

The spilled byte is s6 and is still there. Reading the stack writes directly
settles it: retail saves ten registers and writes sp10 at sp+16, we do exactly
the same plus `sb $a0,0x20($sp)` -- same register budget, one extra byte in
the frame. Four eliminations measured and dead, including s32 s6 itself, which
removes the sb but costs lhu +2 and eight points of alignment.

Build green at 84747e64.

+++ parked/func_80031874.c
@@ -15,7 +15,7 @@ void func_80031874(u8 *arg0, s32 arg1) {
     s32 v;
     u16 fp;
     u8 s6;
-    u8 t;
+    s32 t;
     s32 s2;
 
     g = (u8 *)0x1F800020;
